# docs/swagger.yaml
openapi: 3.0.3
info:
  title: StyleSync Backend
  description: Production REST API & AI Vision Advisor Engine for StyleSync
  version: "1.0"
  contact:
    name: StyleSync AI Engineering
servers:
  - url: http://localhost:5000
    description: Local Backend Server (http://localhost:5000)
  - url: /
    description: Current Host Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT Bearer token generated from /api/v1/auth/login or /api/v1/auth/register
...
paths:
...
  /api/v1/profile:
    get:
      summary: Get User Profile & Physical Traits
      description: Returns user physical traits (faceShape, skinUndertone, colorSeason, bodySilhouette) and styling preferences.
      security:
        - BearerAuth: []
      tags: [profile]
      responses:
        200:
          description: Full profile data
    put:
      summary: Update User Styling Preferences
      description: Update brand preferences, sizes, avoid colors, and budget limits.
      security:
        - BearerAuth: []
      tags: [profile]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                avatar: { type: string }
                physicalTraits: { type: object }
                preferences: { type: object }
      responses:
        200:
          description: Profile updated successfully
...
  /api/v1/outfits:
    get:
      summary: Get Saved Outfits
      description: Returns user saved coordinated outfits.
      security:
        - BearerAuth: []
      tags: [outfit-builder]
      responses:
        200:
          description: List of saved outfits
    post:
      summary: Save Custom Outfit to Lookbook
      description: Save a customized or curated outfit with item IDs to database.
      security:
        - BearerAuth: []
      tags: [outfit-builder]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemIds]
              properties:
                title: { type: string, example: "Smart Friday Ensemble" }
                occasion: { type: string, example: "Smart Casual" }
                weather: { type: string, example: "Mild" }
                style: { type: string, example: "Smart Casual" }
                itemIds:
                  type: array
                  items: { type: string }
      responses:
        201:
          description: Outfit saved successfully
...